Use Base10 by default; I like human-readable JSON.
This starts setting up the int-keyed maps to respect the current NumeralBase setting; more should follow suit. The idea is that if you want a scrambled base, as much should be scrambled as possible. There could also be a mechanic to switch which base this uses based on something complicated behind the scenes.