Sort-of add Junction JSON support.
The class structure of a typical Junction involves at least two classes in a tree shape, which could be hard to handle with a real serializer. I'm currently relying on class tags and automatic serialization to handle Junctions, though for Junction<String> you can just toString() and Junction.parse() it back.